chore: move landing page components into pages directory
