+++ frontend/src/components/pages/landing/LandingHero/OrganizationsSlider.tsx
