+++ frontend/src/components/pages/landing/LandingHero/index.tsx
