+++ frontend/src/components/pages/landing/LandingPromo.tsx
