+++ frontend/src/components/pages/landing/LandingSection.tsx
