+++ frontend/src/components/pages/landing/index.ts
