+++ frontend/src/pages/index.tsx
@@ -1,7 +1,7 @@
 import { NoSsr } from "@mui/material";
 
-import { LandingHero, LandingListings, LandingPromo } from "@/components/landing";
-import { LandingSection } from "@/components/landing/LandingSection";
+import { LandingHero, LandingListings, LandingPromo } from "@/components/pages/landing";
+import { LandingSection } from "@/components/pages/landing/LandingSection";
 import { NextPageWithLayout } from "@/lib/next";
 
 const IndexPage: NextPageWithLayout = () => {
